app/requirements.txt
uvicorn
fastapi
pydantic

aiohttp
aiocache
aiofiles

singleton_package
json-advanced
python-dotenv
debugpy
ipython

beanie
redis

usso>=0.23.1
telegram-bale-bot>=4.19.5
tapsage-client>=0.8.1
metisai>=0.3.3
openai
replicate
apscheduler

apify_client

pandas
google-auth
gspread
gspread-dataframe